Added random events to that text file for later creation in the to be created event manager. Added the max quantities for each item type such as only 20 oxen, 99 bullets, etc. Converted the balance, price, cost, and other financial variables to use floats. Finally for today created all the stores that are located on the trail with their respective prices, the only unique thing amongst all the stores is the cost of the goods so that is what is added to each static read only collection of items that the stores can use as initial inventory seeds.

+++ Src/TrailEntities/Entities/Items/Item.cs
@@ -13,7 +13,7 @@ namespace TrailEntities
         /// <summary>
         ///     Initializes a new instance of the <see cref="T:TrailEntities.Item" /> class.
         /// </summary>
-        protected Item(uint cost, uint quantity)
+        protected Item(float cost, uint quantity)
         {
             _cost = cost;
             _quantity = quantity;
